Struct cobalt::client::Client

pub struct Client {
    // some fields omitted
}

A server client that uses a virtual UDP connection for reliable message transmission.

Methods

impl Client

fn new(config: Config) -> Client

Creates a new client with the given configuration.

fn peer_addr(&self) -> Option<SocketAddr>

Returns the address of the server the client is currently connected to.

fn connect<T: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, address: T) -> Result<(), Error>

Tries to establish connection to the server specified by the address.

The server must use a compatible configuration in order for the connection to be actually established.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

fn close(&mut self)

Closes the clients connections to the server.