cobalt::cobalt_model

Struct DateTime

pub struct DateTime { /* private fields */ }

Liquid’s native date + time type.

Implementations

impl DateTime

pub fn now() -> DateTime

Create a DateTime from the current moment.

pub fn from_ymd(year: i32, month: u8, day: u8) -> DateTime

Makes a new NaiveDate from the calendar date (year, month and day).

Panics on the out-of-range date, invalid month and/or day.

pub fn from_str(other: &str) -> Option<DateTime>

Convert a str to Self

pub fn with_date(self, other: Date) -> DateTime

Replace date with other.

pub fn with_offset(self, offset: UtcOffset) -> DateTime

Changes the associated time zone. This does not change the actual DateTime (but will change the string representation).

pub fn date(self) -> Date

Retrieves a date component.

pub fn format(&self, fmt: &str) -> Result<String, DateFormatError>

Formats the combined date and time with the specified format string.

See the chrono::format::strftime module on the supported escape sequences.

pub fn to_rfc2822(&self) -> String

Returns an RFC 2822 date and time string such as Tue, 1 Jul 2003 10:52:37 +0200.

impl DateTime

pub fn year(&self) -> i32

Get the year of the date.

pub fn month(&self) -> u8

Get the month.

pub fn day(&self) -> u8

Get the day of the month.

pub fn ordinal(&self) -> u16

Get the day of the year.

The returned value will always be in the range 1..=366 (1..=365 for common years).

pub fn iso_week(&self) -> u8

Get the ISO week number.

The returned value will always be in the range 1..=53.