Enum cobalt::ConnectionEvent

pub enum ConnectionEvent {
    Connected,
    FailedToConnect,
    Lost(bool),
    Closed(bool),
    Message(Vec<u8>),
    PacketLost(Vec<u8>),
    CongestionStateChanged(bool),
}

Enum of connection related network events.

Variants
Connected

Emitted once a the connection has been established.

FailedToConnect

Emitted when a connection attempt failed.

Lost(bool)

Emitted when the already established connection is lost.

Closed(bool)

Emitted when the already established connection is closed programmatically.

Message(Vec<u8>)

Emitted for each message that is received over the connection.

PacketLost(Vec<u8>)

Event emitted for each packet which was not confirmed by the remote end of the connection within the specified limits.

CongestionStateChanged(bool)

Emitted each time the connection's congestion state changes.