Enum cobalt::MessageKind

pub enum MessageKind {
    Instant,
    Reliable,
    Ordered,
}

Enum for specification of a message handling algorithm.

Variants
Instant

Message that is going be send exactly once and ignored in case its containing packet is lost. No guarantees are made as for the order in which a message of this kind is going to be received by a remote queue.

Reliable

Message that is going to be re-send in case its containing packet is lost. No guarantees are made as for the order in which a message of this kind is going to be received by a remote queue.

Ordered

Message that is going to be re-send in case its containing packet is lost and is also guaranteed to arrive in-order, meaning that if you send two Ordered messages and the second arrives first in the remote queue , the remote queue will buffer the second message until the first one arrives and then make both of them available to the application at once.