Struct cobalt::server::Server

pub struct Server {
    // some fields omitted
}

A multi-client server that uses a virtual UDP connection for reliable message transmission.

Methods

impl Server

fn new(config: Config) -> Server

Creates a new server with the given configuration.

fn local_addr(&self) -> Option<SocketAddr>

Returns the local address the server is currently bound to.

fn bind<T: ToSocketAddrs>(&mut self, handler: &mut Handler<Server>, address: T) -> Result<(), Error>

Tries to bind the server to the specified local address and actively listens for and manages incoming client connections.

The clients must use a compatible configuration in order for connections to be actually established.

The handler is a struct that implements the Handler trait in order to handle events from the server and its connections.

fn shutdown(&mut self)

Shuts down the server, closing all active connections.