Trait cobalt::shared::handler::Handler

pub trait Handler<T> {
    fn bind(&mut self, _: &mut T) { ... }
    fn tick_connections(&mut self, _: &mut T, _: &mut HashMap<ConnectionID, Connection>) { ... }
    fn shutdown(&mut self, _: &mut T) { ... }
    fn connect(&mut self, _: &mut T) { ... }
    fn tick_connection(&mut self, _: &mut T, _: &mut Connection) { ... }
    fn close(&mut self, _: &mut T) { ... }
    fn connection(&mut self, _: &mut T, _: &mut Connection) { ... }
    fn connection_failed(&mut self, _: &mut T, _: &mut Connection) { ... }
    fn connection_packet_lost(&mut self, _: &mut T, _: &mut Connection, _: &[u8]) { ... }
    fn connection_congested(&mut self, _: &mut T, _: &mut Connection) { ... }
    fn connection_lost(&mut self, _: &mut T, _: &mut Connection) { ... }
}

Trait for implementation of client / server event handling.

Provided Methods

fn bind(&mut self, _: &mut T)

Method that is called once a Server has successfully bound itself to its local address.

fn tick_connections(&mut self, _: &mut T, _: &mut HashMap<ConnectionID, Connection>)

Method that is called each time a Server "ticks". A "tick" occurs in-between the receiving and sending data from / to connections.

fn shutdown(&mut self, _: &mut T)

A method that is called once a Server is going to shutdown.

fn connect(&mut self, _: &mut T)

Method that is called once a Client has successfully bound itself to its local address.

fn tick_connection(&mut self, _: &mut T, _: &mut Connection)

Method that is called each time a Client "ticks". A "tick" occurs in-between the receiving and sending data from / to connections.

fn close(&mut self, _: &mut T)

Method that is called once a Client is going to close.

fn connection(&mut self, _: &mut T, _: &mut Connection)

Method that is called each time a new connection is established.

fn connection_failed(&mut self, _: &mut T, _: &mut Connection)

Method that is called each time a connection fails to establish.

fn connection_packet_lost(&mut self, _: &mut T, _: &mut Connection, _: &[u8])

Method that is called each time a packet send by a connection is lost.

fn connection_congested(&mut self, _: &mut T, _: &mut Connection)

Method that is called each time the congestion state of connection changes.

fn connection_lost(&mut self, _: &mut T, _: &mut Connection)

Method that is called each time a connection is lost and dropped.