Struct cobalt::BinaryRateLimiter

pub struct BinaryRateLimiter {
    // some fields omitted
}

Implementation of a binary state rate limiter for congestion avoidance.

It is based on the example design from the following article: http://gafferongames.com/networking-for-game-programmers/reliability-and-flow-control/

Methods

impl BinaryRateLimiter

fn new(config: &Config) -> Box<BinaryRateLimiter>

Creates a new rate limiter.