Struct cobalt::ClientStream

pub struct ClientStream {
    // some fields omitted
}

Implementation of a stream based Client interface suitable for event polling.

Methods

impl ClientStream

fn new(config: Config) -> ClientStream

Creates a new client stream with the given configuration.

fn from_client(client: Client) -> ClientStream

Consumes the passed in Client instance converting it into a ClientStream.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the address of the server the stream is currently connected to.

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address that the stream is sending from.

fn rtt(&self) -> u32

Returns the average roundtrip time for this stream's underlying connection.

fn packet_loss(&self) -> f32

Returns the percent of packets that were sent and never acknowledged over the total number of packets that have been send across this stream.

fn stats(&self) -> Stats

Returns statistics (i.e. bandwidth usage) for the last second.

fn bytes_sent(&self) -> u32

Returns the number of bytes sent over the last second.

fn bytes_received(&self) -> u32

Returns the number of bytes received over the last second.

fn config(&self) -> Config

Returns a copy of the stream's current configuration.

fn set_config(&mut self, config: Config)

Overrides the stream's current configuration with the one provided.

fn connect<A: ToSocketAddrs>(&mut self, addr: A) -> Result<(), Error>

Establishes a connection with the server at the specified address.

fn receive(&mut self) -> Result<ClientEvent, TryRecvError>

Receives the next incoming message from the stream's underlying connection.

fn send(&mut self, kind: MessageKind, payload: Vec<u8>) -> Result<(), Error>

Queues a message of the specified kind along with its payload to be send with the next flush call.

fn flush(&mut self) -> Result<(), Error>

Sends all queued messages over the stream's underlying connection.

fn reset(&mut self) -> Result<(), Error>

Resets the stream's underlying connection to the server.

fn close(&mut self) -> Result<(), Error>

Closes the stream's underlying connection to the server.