Crate cobalt

cobalt is a networking library which provides virtual connections over UDP for both unreliable, reliable and optional in-order delivery of messages.

It is primarily designed to be used as the basis of real-time, low latency, single server / multi client systems.

The library provides the underlying architecture required for handling and maintaining virtual connections over UDP sockets and takes care of sending raw messages over the established client-server connections with minimal overhead.

cobalt is also fully configurable and can be integrated in a variety of ways, as can be seen in the examples below.

Getting started

cobalt supports a number of different integration strategies, the most forward one being the use of so called handlers. The Handler trait acts as a event proxy for both server and client events that are emitted from the underlying tick loop inside the library.

Handler based integration

The handler based approach suits itself especially well for the server side where all updates on any connections will be performed inside a single tick handler.

use std::collections::HashMap;
use cobalt::{Config, Connection, ConnectionID, Handler, Server, MessageKind};

struct GameServer {
    tick: u8
}
impl Handler<Server> for GameServer {

    fn bind(&mut self, server: &mut Server) {
        // Since this is a runnable doc, we'll just exit right away
        server.shutdown().unwrap();
    }

    fn tick_connections(
        &mut self, _: &mut Server,
        connections: &mut HashMap<ConnectionID, Connection>
    ) {

        for (_, conn) in connections.iter_mut() {
            // Receive player input
            for msg in conn.received() {
                println!("Received message from client: {:?}", msg);
            }
        }

        // Advance game state
        self.tick = self.tick.wrapping_add(1);

        // Send state updates to players
        for (_, conn) in connections.iter_mut() {
            conn.send(MessageKind::Instant, b"Hello World".to_vec());
        }

    }

    fn shutdown(&mut self, _: &mut Server) {
        // Logging and things
    }

    fn connection(&mut self, _: &mut Server, _: &mut Connection) {
        // Create Player, send MOTD etc.
    }

    fn connection_lost(&mut self, _: &mut Server, _: &mut Connection) {
        // Remove Player
    }

}

let mut handler = GameServer {
    tick: 0
};
let mut server = Server::new(Config::default());
server.bind(&mut handler, "127.0.0.1:7156").ok();

The client version of the handler based approach looks almost identical - except for a few methods sporting different names.

This method is best used when separate logic / rendering threads are used.

use cobalt::{Config, Connection, Handler, Client, MessageKind};

struct GameClient {
    tick: u8
};

impl Handler<Client> for GameClient {

    fn connect(&mut self, client: &mut Client) {
        // Since this is a runnable doc, we'll just exit right away
        client.close().unwrap();
    }

    fn tick_connection(&mut self, _: &mut Client, conn: &mut Connection) {

        // Receive incoming state from the server
        for msg in conn.received() {
            println!("Received message {:?}", msg);
        }

        // Advance game state
        self.tick = self.tick.wrapping_add(1);

        // Send some message to server
        conn.send(MessageKind::Instant, b"Hello World".to_vec());

    }

    fn close(&mut self, _: &mut Client) {
        // Exit game
    }

    fn connection(&mut self, _: &mut Client, _: &mut Connection) {
        // Request map list and settings from server
    }

    fn connection_failed(&mut self, _: &mut Client, _: &mut Connection) {
        // Failed to connect to the server
    }

    fn connection_lost(&mut self, _: &mut Client, _: &mut Connection) {
        // Inform the user
    }

}

let mut handler = GameClient {
    tick: 0
};
let mut client = Client::new(Config::default());
client.connect(&mut handler, "127.0.0.1:7156").ok();

Stream based integration

For situations where a event loop is already in place (e.g. game engines), a stream based approach can be used in order to achieve easier integration in single threaded scenarios.

use cobalt::{ClientEvent, ClientStream, Config, MessageKind};

// Create a new stream
let mut stream = ClientStream::new(Config {
    send_rate: 15,
    .. Default::default()
});

// Initiate the connection
stream.connect("127.0.0.1:5555").ok();

// Inside of the existing event loop
// loop {

    // Receive incoming messages
    while let Ok(event) = stream.receive() {
        match event {
            ClientEvent::Connection => println!("Connection established"),
            ClientEvent::ConnectionLost => println!("Connection lost"),
            ClientEvent::Tick => {
                // Handle network related logic, advance actual in game time
            },
            ClientEvent::Message(payload) => println!("Received message: {:?}", payload),
            _ => {}
        }
    }

    // Send some messages
    stream.send(MessageKind::Instant, b"Hello World".to_vec()).ok();

    // Send outgoing messages
    stream.flush().ok();

// }

Low level synchronous integration

For times when even the stream based abstraction is too much, there's also the option to use the underlying synchronous client implement on top of which both the handler and stream based approaches are built.

use cobalt::{Client, Config, Handler, MessageKind};

struct SyncHandler;
impl Handler<Client> for SyncHandler {}

let mut handler = SyncHandler;
let mut client = Client::new(Config::default());
let mut state = client.connect_sync(
    &mut handler,
    "127.0.0.1:7156"

).unwrap();

// Receive from the server
client.receive_sync(&mut handler, &mut state, 0);

// Tick the connection
client.tick_sync(&mut handler, &mut state);

// Send a message
state.send(MessageKind::Instant, b"Hello World".to_vec());

// Send any pending messages via the connection
client.send_sync(&mut handler, &mut state);

Structs

BinaryRateLimiter
Implementation of a binary state rate limiter for congestion avoidance.

Client
Implementation of a single-server client with handler based event dispatch.

ClientState
A structure used for synchronous calls on a Client instance.

ClientStream
Implementation of a stream based Client interface suitable for event polling.

Config
Structure defining connection and message configuration options.

Connection
Implementation of a reliable, virtual connection logic.

ConnectionID
Representation of a random ID for connection identification purposes.

Server
Implementation of a multi-client server with handler based event dispatch.

Stats
A structure containing stats data average of the course of one second.

UdpSocket
Non-blocking abstraction over a UDP socket.

Enums

ClientEvent
Enum of stream related network events.

ConnectionState
Enum indicating the state of a connection.

MessageKind
Enum for specification of a message handling algorithm.

Traits

Handler
Trait for implementation of a client / server event proxy.

RateLimiter
Trait for implementation of a network congestion avoidance algorithm.

Socket
Trait for implementation of a non-blocking UDP socket.

Type Definitions

ConnectionMap
Type alias for connection mappings.