Struct cobalt::shared::Config

pub struct Config {
    pub send_rate: u32,
    pub packet_max_size: usize,
    pub protocol_header: [u8; 4],
    pub packet_drop_threshold: u32,
    pub connection_init_threshold: u32,
    pub connection_drop_threshold: u32,
    pub message_quota_instant: f32,
    pub message_quota_reliable: f32,
    pub message_quota_ordered: f32,
}

A generic configuration object.

Fields

send_rate

Number of packets send per second. Default is 30.

packet_max_size

Maximum bytes that can be received / send in one packet. Default 1400.

protocol_header

32-Bit Protocol ID used to identify UDP related packets. Default is [1, 2, 3, 4].

packet_drop_threshold

Maximum RTT in milliseconds before a packet is considered lost. Default is 1000.

connection_init_threshold

Maximum time in milliseconds until the first packet must be received before a connection attempt fails. Default is 100.

connection_drop_threshold

Maximum time in milliseconds between any two packets before the connection gets dropped. Default is 1000.

message_quota_instant

The percent of available packet bytes to use when serializing MessageKind::Instant into a packet via a MessageQueue.

message_quota_reliable

The percent of available packet bytes to use when serializing MessageKind::Reliable into a packet via a MessageQueue.

message_quota_ordered

The percent of available packet bytes to use when serializing MessageKind::Ordered into a packet via a MessageQueue.