Struct cobalt::Client

pub struct Client {
    // some fields omitted
}

Implementation of a single-server client with handler based event dispatch.

Methods

impl Client

fn new(config: Config) -> Client

Creates a new client with the given configuration.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the address of the server the client is currently connected to.

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address that the client is sending from.

fn connect<A: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, addr: A) -> Result<(), Error>

Establishes a connection with the server at the specified address by creating a local socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

fn connect_from_socket<S: Socket, A: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, addr: A, socket: S) -> Result<(), Error>

Establishes a connection with the server at the specified address by using the specified socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

fn close(&mut self) -> Result<(), Error>

Closes the connection to the server.

This exits the tick loop, resets the connection and shuts down the underlying socket the client was sending from.