Struct cobalt::Client

pub struct Client {
    // some fields omitted
}

Implementation of a single-server client with handler based event dispatch.

There are two ways of creating and connection a client instance:

Blocking mode to be used in it dedicated thread, with asynchronous Handler callbacks, available via the methods without the sync postfix.

Non-Blocking mode, with synchronous Handler callbacks, available via the methods with the sync postfix.

Methods

impl Client

fn new(config: Config) -> Client

Creates a new client with the given configuration.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the address of the server the client is currently connected to.

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address that the client is sending from.

fn stats(&mut self) -> Stats

Returns statistics (i.e. bandwidth usage) for the last second.

fn connect<A: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, addr: A) -> Result<(), Error>

Establishes a connection with the server at the specified address by creating a local socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

This method starts the tick loop, blocking the calling thread.

fn connect_from_socket<S: Socket, A: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, addr: A, socket: S) -> Result<(), Error>

Establishes a connection with the server at the specified address by using the specified socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

This method starts the tick loop, blocking the calling thread.

fn close(&mut self) -> Result<(), Error>

Asynchronously closes the connection to the server.

This exits the tick loop, resets the connection and shuts down the underlying socket the client was sending and receiving from.

fn connect_sync<A: ToSocketAddrs>(&mut self, handler: &mut Handler<Client>, addr: A) -> Result<ClientState<UdpSocket>, Error>

Establishes a connection with the server at the specified address by creating a local socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

This method returns a ClientState instance for this client, which can be used with other synchronous Client methods.

fn connect_from_socket_sync<A: ToSocketAddrs, S: Socket>(&mut self, handler: &mut Handler<Client>, addr: A, socket: S) -> Result<ClientState<S>, Error>

Establishes a connection with the server at the specified address by using the specified socket for message sending.

The server must use a compatible connection / packet configuration.

The handler is a struct that implements the Handler trait in order to handle events from the client and its connection.

This method returns a ClientState instance for this client, which can be used with other synchronous Client methods.

fn receive_sync<S: Socket>(&mut self, handler: &mut Handler<Client>, state: &mut ClientState<S>, tick_delay: u32)

Receives all currently buffered incoming packet from the underlying connection.

fn tick_sync<S: Socket>(&mut self, handler: &mut Handler<Client>, state: &mut ClientState<S>)

Performs exactly on tick of the underlying connection.

fn send_sync<S: Socket>(&mut self, handler: &mut Handler<Client>, state: &mut ClientState<S>)

Sends exactly on outgoing packet from the underlying connection.

fn close_sync<S: Socket>(&mut self, handler: &mut Handler<Client>, state: &mut ClientState<S>) -> Result<(), Error>

Closes the connection to the server.

This resets the connection and shuts down the underlying socket the client was sending and receiving from.