Struct cobalt::Server

pub struct Server {
    // some fields omitted
}

Implementation of a multi-client server with handler based event dispatch.

Methods

impl Server

fn new(config: Config) -> Server

Creates a new server with the given configuration.

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address that the server is bound to.

fn stats(&mut self) -> Stats

Returns statistics (i.e. bandwidth usage) for the last second.

fn bind<A: ToSocketAddrs>(&mut self, handler: &mut Handler<Server>, addr: A) -> Result<(), Error>

Binds the server to the specified local address by creating a socket and actively listens for incoming client connections.

Clients connecting to the server must use a compatible connection / packet configuration in order to be able to connect.

The handler is a struct that implements the Handler trait in order to handle events from the server and its connections.

fn bind_to_socket<S: Socket>(&mut self, handler: &mut Handler<Server>, socket: S) -> Result<(), Error>

Binds the server to specified socket and actively listens for incoming client connections.

Clients connecting to the server must use a compatible connection / packet configuration in order to be able to connect.

The handler is a struct that implements the Handler trait in order to handle events from the server and its connections.

fn shutdown(&mut self) -> Result<(), Error>

Shuts down the server, closing all active client connections.

This exits the tick loop, resets all connections and shuts down the underlying socket the server was bound to.